move AOVMenuBox class's code to it's base class. Now, function UOVInterface::ShowDlgJson() is compatible with function UOVInterface::onEventSystemMenu()
